update: cálculo de la progresión de color en CountdownViewModel

+++ app/src/main/java/com/espiralsoft/bisiestus/presentation/states/CountdownUiState.kt
@@ -1,11 +1,11 @@
 package com.espiralsoft.bisiestus.presentation.states
 
 data class CountdownUiState(
-    val weeks: Long = 0,
-    val days: Long = 0,
-    val hours: Long = 0,
-    val minutes: Long = 0,
-    val seconds: Long = 0,
+    val weeks: Int = 0,
+    val days: Int = 0,
+    val hours: Int = 0,
+    val minutes: Int = 0,
+    val seconds: Int = 0,
     val isFeb29: Boolean = false,
     val progressionColor: Float = 0f
 )
